pyproject.toml
[tool.poetry]
name = "butterflymx-444"
version = "0.0.1"
description = ""
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "butterflymx_444"}]

[tool.poetry.dependencies]
python = "^3.11"
butterflymx = "^0.1.1"
python-dotenv = "^1.0.0"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pydantic = "^1.10.7"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.265"
mypy = "^1.2.0"
types-python-jose = "^3.3.4.7"

[tool.poetry.scripts]
create_user = { callable = "butterflymx_444.scripts:create_user" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy"]
files = ["butterflymx_444/**/*.py"]
python_version = "3.10"
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true

[tool.ruff]
line-length = 100
